\name{SymmetricMA}
\alias{SymmetricMA}
\title{ Make a series smooth using moving average }
\description{
It makes the data smooth using moving average of specified window size.  
}
\usage{
SymmetricMA(x, w = 5)
}
\arguments{
  \item{x}{Array-CGH data }
  \item{w}{Window size.}
}

\value{
Smoothed version of original data.
}



\author{M. Shahidul Islam}


\keyword{nonparametric}
